Integration with GRI's events
Via these new events, HLQ has the option to choose whether or not to intervene in the Lord Protector's takeover of GRI and save Grover (if they have gone integralist), or simply say "nah bro". Localization for the choice event is rough, at the moment "Heroes Hesitate" is the no option, while "Who will save the world" is the yes option. Props if you get the song reference.

This update also includes an updated flag for Hellquill alongside of fresh flags for Nige- I mean uh, Neu Griffonischess Reich.