Migrating HLQ cosmetics to decision & fixes
This update migrates HLQ's cosmetic changes to the decision to reform the Griffonian empire. It also changes the name of the cosmetic tag to fit more closely with similar cosmetic tags for the decision.
Also contains a fix for missing localization for the GRI version of the Saving Grover event.